Extracted assigment walker
