+++ lib/Core/Inference/FrameWalker.php
@@ -3,11 +3,10 @@
 namespace Phpactor\WorseReflection\Core\Inference;
 
 use Microsoft\PhpParser\Node;
-use Phpactor\WorseReflection\Core\Inference\FrameBuilder\Frame;
 
 interface FrameWalker
 {
     public function canWalk(Node $node);
 
-    public function walk(Frame $frame, Node $node);
+    public function walk(FrameBuilder $builder, Frame $frame, Node $node);
 }
